requirements.txt
Flask==2.3.3
Flask-Babel==3.1.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
Babel==2.12.1
pytz==2023.3
reportlab==4.0.4
Pillow==10.0.0
requests==2.31.0
python-dotenv==1.0.0
openai==1.3.0
anthropic==0.5.0
google-generativeai==0.3.0
pydantic==2.4.2
typing-extensions==4.7.1
colorama==0.4.6
tqdm==4.66.1 